Adds dynamix z-value updating.
As new layers are added, z-values must be updated to properly layer models. This commit allows "infinite" layering.

+++ src/lib/ren2/webgl/shader/source/vertex.ts
@@ -3,8 +3,8 @@ const vertexShaderSource = `
   attribute vec2 aTextureCoordinate;
 
   // uint is not supported in WebGL (GLSL 100) - it is in WebGL 2 (GLSL 300)
-  uniform  int uLayer;
-  uniform vec2 uLocation;
+  uniform float uLayer;
+  uniform  vec2 uLocation;
 
   varying highp vec2 vTextureCoordinate;
